add cmake

+++ src/aybabtu/base64.cpp
@@ -65,10 +65,10 @@ std::vector<uint8_t> base64::decode(const std::string& encoded_data)
     uint8_t b[4];
     for (uint32_t i = 0U; i < encoded_data.length(); i += 4)
     {
-        b[0] = base64_chars.find(encoded_data[i]);
-        b[1] = base64_chars.find(encoded_data[i + 1]);
-        b[2] = base64_chars.find(encoded_data[i + 2]);
-        b[3] = base64_chars.find(encoded_data[i + 3]);
+        b[0] = (uint8_t)base64_chars.find(encoded_data[i]);
+        b[1] = (uint8_t)base64_chars.find(encoded_data[i + 1]);
+        b[2] = (uint8_t)base64_chars.find(encoded_data[i + 2]);
+        b[3] = (uint8_t)base64_chars.find(encoded_data[i + 3]);
         result.push_back((b[0] << 2) | (b[1] >> 4));
         if (b[2] < 64)
         {
